Adding Meta Tag Generater.
This improves the global handling of Meta Tags throughout the system
via simplified directory structure.

The Class "Meta" is now available to be used in Blade.
The methods have not been implemented yet. so there might be a problem
in using the Meta class on the blades without methods defined manually.

Signed-off-by: Gnanakeethan Balasubramaniam <gnanakeethan@gmail.com>